Move react controls to datascience-ui directory in prep of bundling
