+++ src/datascience-ui/history-react/collapseButton.css
