+++ src/datascience-ui/history-react/collapseButton.tsx
