+++ src/datascience-ui/history-react/images.d.ts
