+++ src/datascience-ui/history-react/menuBar.tsx
