+++ src/datascience-ui/history-react/registerServiceWorker.ts
