+++ src/datascience-ui/history-react/transforms.ts
