+++ src/datascience-ui/react-common/locReactSide.ts
