+++ src/datascience-ui/react-common/relativeImage.tsx
