+++ src/datascience-ui/react-common/themeDetector.ts
