+++ src/test/datascience/history.unit.test.tsx
@@ -1,7 +1,32 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
 'use strict';
+
+import * as assert from 'assert';
+import { mount } from 'enzyme';
+import * as React from 'react';
+import * as TypeMoq from 'typemoq';
+import { Disposable } from 'vscode';
+
+import {
+    IWebPanel,
+    IWebPanelMessage,
+    IWebPanelMessageListener,
+    IWebPanelProvider
+} from '../../client/common/application/types';
+import { PlatformService } from '../../client/common/platform/platformService';
+import { IFileSystem, IPlatformService } from '../../client/common/platform/types';
+import { IPythonExecutionFactory, IPythonExecutionService } from '../../client/common/process/types';
+import { IDisposableRegistry, ILogger } from '../../client/common/types';
+import { HistoryProvider } from '../../client/datascience/historyProvider';
+import { JupyterServerProvider } from '../../client/datascience/jupyterServerProvider';
+import { IHistoryProvider, IJupyterServerProvider } from '../../client/datascience/types';
+import { IServiceContainer } from '../../client/ioc/types';
+import { IVsCodeApi } from '../../datascience-ui//react-common/postOffice';
+import { Cell } from '../../datascience-ui/history-react/cell';
+import { MainPanel } from '../../datascience-ui/history-react/mainPanel';
+import { MockPythonExecutionService } from './executionServiceMock';
+import { waitForUpdate } from './reactHelpers';
 
 // Custom module loader so we skip .css files that break non webpack wrapped compiles
 // tslint:disable-next-line:no-var-requires no-require-imports
@@ -22,26 +47,6 @@ const Module = require('module');
         return _require(this, path);
     };
 })();
-
-import * as assert from 'assert';
-import { mount } from 'enzyme';
-import * as React from 'react';
-import * as TypeMoq from 'typemoq';
-import { Disposable } from 'vscode';
-import { IWebPanel, IWebPanelMessage, IWebPanelMessageListener, IWebPanelProvider  } from '../../client/common/application/types';
-import { PlatformService } from '../../client/common/platform/platformService';
-import { IFileSystem, IPlatformService } from '../../client/common/platform/types';
-import { IPythonExecutionFactory, IPythonExecutionService } from '../../client/common/process/types';
-import { IDisposableRegistry, ILogger } from '../../client/common/types';
-import { Cell } from '../../client/datascience/history-react/cell';
-import { MainPanel } from '../../client/datascience/history-react/mainPanel';
-import { HistoryProvider } from '../../client/datascience/historyProvider';
-import { JupyterServerProvider } from '../../client/datascience/jupyterServerProvider';
-import { IVsCodeApi } from '../../client/datascience/react-common/postOffice';
-import { IHistoryProvider, IJupyterServerProvider  } from '../../client/datascience/types';
-import { IServiceContainer } from '../../client/ioc/types';
-import { MockPythonExecutionService } from './executionServiceMock';
-import { waitForUpdate } from './reactHelpers';
 
 // tslint:disable-next-line:max-func-body-length
 suite('History output tests', () => {
